requirements.txt
python-telegram-bot==22.3
python-dotenv==1.0.0
requests==2.31.0
httpx
